relocated tests and created tests for elevator
