+++ src/test/ElevatorDoorTest.java
@@ -1,0 +1,27 @@
+package test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+import org.junit.jupiter.api.Test;
+import elevator.ElevatorDoor;
+
+class ElevatorDoorTest {
+	
+	@Test
+	void testDoorOpening() throws InterruptedException {
+		//Create a Elevator Door
+		ElevatorDoor door = new ElevatorDoor();
+
+		//verify default status
+		assertEquals(false, door.doorIsOpen());
+
+		//verify openDoor method
+		door.openDoor();
+		assertEquals(true, door.doorIsOpen());
+
+		//verify closeDoor method
+		door.closeDoor();
+		assertEquals(false, door.doorIsOpen());
+
+	}
+}
